test/problem/facebook/fb_1/ValidPalindrome3Test.kt
package problem.facebook.fb_1

import org.junit.Assert
import org.junit.Test

/**
 * Created by randheercode
 * Date: 18/11/20
 * Time: 8:05 am
 */
class ValidPalindrome3Test {

    private val obj = ValidPalindrome3()

    @Test
    fun isValidPalindrome() {
        Assert.assertTrue(obj.isValidPalindrome("abcdeca", 2))
        Assert.assertTrue(obj.isValidPalindrome("fcgihcgeadfehgiabegbiahbeadbiafgcfchbcacedbificicihibaeehbffeidiaiighceegbfdggggcfaiibefbgeegbcgeadcfdfegfghebcfceiabiagehhibiheddbcgdebdcfegaiahibcfhheggbheebfdahgcfcahafecfehgcgdabbghddeadecidicchfgicbdbecibddfcgbiadiffcifiggigdeedbiiihfgehhdegcaffaggiidiifgfigfiaiicadceefbhicfhbcachacaeiefdcchegfbifhaeafdehicfgbecahidgdagigbhiffhcccdhfdbd", 216))
    }

}